requirements.txt
# Google Workspace MCP Server Dependencies

# Core MCP and FastAPI
fastapi>=0.115.12
fastmcp>=2.3.3
mcp[cli]>=1.6.0
uvicorn>=0.34.2
sse-starlette>=2.3.3

# Google API dependencies
google-api-python-client>=2.168.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.2

# HTTP and networking
httpx>=0.28.1
requests>=2.31.0

# Security and authentication
pyjwt>=2.10.1
cryptography>=41.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.1.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
